Factor out source creation tests
